requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
aniso8601==9.0.1
certifi==2021.5.30
chardet==4.0.0
click==7.1.1
cycler==0.10.0
efficientnet-pytorch==0.6.3
enum34==1.1.10
Flask==1.0
Flask-RESTful==0.3.9
future==0.18.2
idna==2.10
itsdangerous==0.24
Jinja2==2.11.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.4
opencv-python==4.2.0.34
pandas==1.0.3
piexif==1.1.3
Pillow==8.1.1
pyparsing==2.4.7
PyQt5==5.15.0
PyQt5-sip==12.9.0
PyQtWebEngine==5.15.2
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2020.1
PyYAML==5.4
requests==2.25.1
six==1.14.0
torch==1.6.0+cpu
torchvision==0.7.0+cpu
urllib3==1.26.6
Werkzeug==0.15.5
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
gunicorn